Add deletePointEntry, update styling

+++ src/VolunteerDetails.jsx
@@ -1,6 +1,6 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import PointEntriesForm from './PointEntriesForm';
-import { Link, useNavigate, useOutletContext } from 'react-router-dom';
+import { useOutletContext } from 'react-router-dom';
 
 function VolunteerDetails({ volunteerDetail, showDetails, setShowDetails }) {
   const [loggedIn, setLoggedIn] = useOutletContext();
@@ -29,7 +29,16 @@ function VolunteerDetails({ volunteerDetail, showDetails, setShowDetails }) {
                     <td>{entry.date}</td>
                     <td>{entry.event}</td>
                     <td>{entry.volunteerRole}</td>
-                    <td>{entry.points}</td>
+                    <td>
+                      <div className="pointsField">
+                        {entry.points}
+                        {loggedIn && (
+                          <span onClick={() => deletePointEntry(index)}>
+                            ✖️
+                          </span>
+                        )}{' '}
+                      </div>
+                    </td>
                   </tr>
                 );
               })
@@ -53,13 +62,34 @@ function VolunteerDetails({ volunteerDetail, showDetails, setShowDetails }) {
     if (e.target.className === 'closeButton' || e.target.className === 'modal')
       setShowDetails(-1);
   }
-  const navigate = useNavigate();
 
   function deleteVolunteer() {
-    fetch(`http://localhost:3000/users/${id}`, {
-      method: 'DELETE',
-    });
-    window.location.href = '/volunteers';
+    if (
+      confirm('Are you sure you want to permanently delete this volunteer?')
+    ) {
+      fetch(`http://localhost:3000/users/${id}`, {
+        method: 'DELETE',
+      });
+      window.location.href = '/volunteers';
+    }
+  }
+
+  function deletePointEntry(deleteIndex) {
+    if (confirm('Are you sure you want to permanently delete this entry?')) {
+      fetch(`http://localhost:3000/users/${id}`, {
+        method: 'PATCH',
+        headers: {
+          'Content-Type': 'Application/JSON',
+        },
+        body: JSON.stringify({
+          pointEntries: volunteerDetail.pointEntries.filter((entry) => {
+            return volunteerDetail.pointEntries.indexOf(entry) !== deleteIndex;
+          }),
+        }),
+      })
+        .then((res) => res.json())
+        .then((updated) => setPointEntries(updated.pointEntries));
+    }
   }
 
   return (
